#37 rename gslab_make to source and update test paths
