+++ source/move_sources.py
@@ -11,11 +11,11 @@ from termcolor import colored
 import colorama
 colorama.init()
 
-import gslab_make.private.metadata as metadata
-from gslab_make.private.exceptionclasses import ColoredError
-from gslab_make.private.movedirective import MoveList
-from gslab_make.private.utility import get_path, format_message
-from gslab_make.write_logs import write_to_makelog
+import source.private.metadata as metadata
+from source.private.exceptionclasses import ColoredError
+from source.private.movedirective import MoveList
+from source.private.utility import get_path, format_message
+from source.write_logs import write_to_makelog
 
 
 def _create_links(paths,
