+++ source/private/movedirective.py
@@ -12,14 +12,14 @@ import subprocess
 from itertools import chain
 
 if (sys.version_info < (3, 0)) and (os.name == 'nt'):
-    import gslab_make.private.subprocess_fix as subprocess_fix
+    import source.private.subprocess_fix as subprocess_fix
 else:
     import subprocess as subprocess_fix
     
-import gslab_make.private.messages as messages
-import gslab_make.private.metadata as metadata
-from gslab_make.private.exceptionclasses import CritError
-from gslab_make.private.utility import convert_to_list, norm_path, file_to_array, format_traceback, decode
+import source.private.messages as messages
+import source.private.metadata as metadata
+from source.private.exceptionclasses import CritError
+from source.private.utility import convert_to_list, norm_path, file_to_array, format_traceback, decode
 
 
 class MoveDirective(object):
