+++ test/test_move_sources.py
@@ -10,11 +10,11 @@ import shutil
 import unittest
 from test.utility import no_stderrout, redirect_stdout, read_file
 
-import gslab_make.private.metadata as metadata
-from gslab_make import start_makelog, remove_dir, clear_dir
-from gslab_make.private.exceptionclasses import CritError, ProgramError
+import source.private.metadata as metadata
+from source import start_makelog, remove_dir, clear_dir
+from source.private.exceptionclasses import CritError, ProgramError
     
-from gslab_make import link_inputs, copy_inputs, link_externals, copy_externals
+from source import link_inputs, copy_inputs, link_externals, copy_externals
 
 class TestLinkInputs(unittest.TestCase):
